Reorganize ntlm tests
